+ Updated SDL2 to v2.0.7 so mouse warping works correctly on windows 10 + console cheat infiniteammo added + moses cheat renamed to god, and cheat persists from map to map + Crouch button added to move down when flying and in water + Jump now moves up when flying and in water + Can hide the HUD with console command g_showhud + Added Anisotropic Filtering to the Graphics menu + Added Texture Filtering as a console command: gl_filter + Added option to remove the camera rolling when moving the mouse on the x axis in the Gamplay menu + Added option to remove the view bobbing in the Gameplay menu + Updated to version v1.1 and displayed it in console.

+++ kex3_anubis/source/system/sdl/systemSDL.cpp
@@ -413,7 +413,7 @@ void kexSystemSDL::SetWindowTitle(const char *string)
 
 void kexSystemSDL::SetWindowGrab(const bool bEnable)
 {
-    SDL_SetWindowGrab(window, (SDL_bool)bEnable);
+	SDL_SetWindowGrab(window, (SDL_bool)bEnable);
 }
 
 //
@@ -422,9 +422,7 @@ void kexSystemSDL::SetWindowGrab(const bool bEnable)
 
 void kexSystemSDL::WarpMouseToCenter(void)
 {
-    SDL_WarpMouseInWindow(window,
-                          (unsigned short)(kex::cSystem->VideoWidth()/2),
-                          (unsigned short)(kex::cSystem->VideoHeight()/2));
+    SDL_WarpMouseInWindow(window, (kex::cSystem->VideoWidth()/2), (kex::cSystem->VideoHeight()/2));
 }
 
 //
